Create CollapsedStoryPoints component
