+++ app/assets/javascripts/components/story/CollapsedStory/CollapsedStoryPoints.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const CollapsedStoryPoints = () => (
+  <div>
+    <span className="Story__estimate">1</span>
+    <span className="Story__estimate">2</span>
+    <span className="Story__estimate">3</span>
+    <span className="Story__estimate">5</span>
+    <span className="Story__estimate">8</span>
+  </div>
+);
+
+export default CollapsedStoryPoints;
